cv/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
# flask>=2.3.0         # Optional: for cloud API endpoint